Rename portal/version components
