--- src/components/main/PortalD/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "PortalD",
-  "version": "0.0.0",
-  "private": true,
-  "main": "PortalD.js"
-}
