--- src/components/main/PortalE/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "PortalE",
-  "version": "0.0.0",
-  "private": true,
-  "main": "PortalE.js"
-}
