+++ src/components/main/VersionA/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "VersionA",
+  "version": "0.0.0",
+  "private": true,
+  "main": "VersionA.js"
+}
