+++ src/components/main/VersionB/VersionB.js
