+++ src/components/main/VersionB/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "VersionB",
+  "version": "0.0.0",
+  "private": true,
+  "main": "VersionB.js"
+}
